@@ -2,7 +2,7 @@ name: charity_discount
 description: Empowering charity.
 publish_to: none
 
-version: 1.12.1+700
+version: 1.13.0+700
 
 environment:
   sdk: '>=2.3.0 <3.0.0'
@@ -12,64 +12,58 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_cupertino_localizations: ^1.0.1
-  firebase_core: ^0.5.3
-  google_sign_in: ^4.5.6
-  firebase_auth: ^0.18.4
-  firebase_analytics: ^6.3.0
-  cloud_firestore: ^0.14.4
-  rxdart: ^0.25.0
-  http: ^0.12.2
-  flushbar:
-    git:
-      url: https://github.com/AndreHaueisen/flushbar
-      ref: master
-  shared_preferences: ^0.5.12
-  easy_localization: ^2.3.4-nullsafety.1
-  flutter_screenutil: ^4.0.2+1
-  url_launcher: ^5.7.10
-  scoped_model: ^2.0.0-nullsafety.0
-  timeline_list: ^0.0.5
-  cached_network_image: ^2.5.0
-  intl: ^0.16.1
-  firebase_remote_config: ^0.4.3
+  another_flushbar: ^1.10.18
+  apple_sign_in: ^0.1.0
+  cached_network_image: ^3.0.0
+  carousel_slider: ^3.0.0
+  charts_flutter: ^0.10.0
+  cloud_firestore: ^1.0.5
   cupertino_icons: ^1.0.2
-  flutter_svg: ^0.19.2+1
-  iban_form_field: ^0.2.1
+  device_info: ^2.0.0-nullsafety.1
+  devicelocale: ^0.4.1
+  easy_localization: ^3.0.0
+  firebase_analytics: ^8.0.0
+  firebase_auth: ^1.1.0
+  firebase_core: ^1.0.3
+  firebase_crashlytics: ^2.0.0
+  firebase_dynamic_links: ^2.0.0
+  firebase_messaging: ^9.1.1
+  firebase_remote_config: ^0.10.0-dev.0
+  firebase_storage: ^8.0.3
+  flutter_cupertino_localizations: ^1.0.1
+  flutter_facebook_login: ^3.0.0
   flutter_html: ^1.1.1
   flutter_rating_bar: ^4.0.0-nullsafety.0
-  flutter_facebook_login: ^3.0.0
-  local_auth: ^0.6.3
-  firebase_messaging: '^8.0.0-dev.11'
-  devicelocale: ^0.3.3
-  device_info: ^2.0.0-nullsafety.1
-  apple_sign_in: ^0.1.0
-  flutter_signin_button: ^1.1.0
-  firebase_storage: '^5.2.0'
-  image_picker: ^0.6.7+17
-  image_cropper: ^1.3.1
-  carousel_slider: ^2.3.1
-  json_annotation: ^3.1.1
-  firebase_crashlytics: ^0.2.4
-  substring_highlight: ^0.1.2
+  flutter_screenutil: ^5.0.0
+  flutter_signin_button: ^2.0.0
+  flutter_svg: ^0.20.0-nullsafety.3
   fluttertoast: ^8.0.0-nullsafety.0
-  firebase_dynamic_links: ^0.6.3
-  share: ^0.6.5+4
-  package_info: ^0.4.3+2
-  charts_flutter: ^0.9.0
-  timeline_tile: ^1.0.0
   get_it: ^6.0.0-nullsafety.1
-  percent_indicator: ^2.1.9
-
-dependency_overrides:
-   intl: ^0.17.0-nullsafety.2
+  google_sign_in: ^4.5.6
+  http: ^0.13.1
+  iban_form_field: ^0.2.1
+  image_cropper: ^1.3.1
+  image_picker: ^0.7.4
+  intl: ^0.17.0
+  json_annotation: ^4.0.1
+  local_auth: ^1.1.4
+  package_info: ^2.0.0
+  percent_indicator: ^3.0.1
+  rxdart: ^0.26.0
+  scoped_model: ^2.0.0-nullsafety.0
+  share: ^2.0.1
+  shared_preferences: ^2.0.5
+  substring_highlight: ^0.1.2
+  timeline_list: ^0.0.5
+  timeline_tile: ^2.0.0
+  url_launcher: ^6.0.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.8.1
+  flutter_launcher_icons: ^0.9.0
   build_runner: ^1.10.2
-  json_serializable: ^3.5.0
+  json_serializable: ^4.1.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
