@@ -2,13 +2,13 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 
 Widget reAuthDialogBuilder(context) {
-  Widget cancelButton = FlatButton(
+  Widget cancelButton = TextButton(
     child: Text(tr('cancel')),
     onPressed: () {
       Navigator.pop(context, false);
     },
   );
-  Widget continueButton = FlatButton(
+  Widget continueButton = TextButton(
     child: Text('OK'),
     onPressed: () {
       Navigator.pop(context, true);
