@@ -39,7 +39,8 @@ class _ProductsScreenState extends State<ProductsScreen> {
     _productsScrollController = ScrollController();
     _productsScrollController.addListener(() {
       if (_searchInProgress == false && _products.length < _totalProducts) {
-        if (_productsScrollController.position.pixels > 0.9 * _productsScrollController.position.maxScrollExtent) {
+        if (_productsScrollController.position.pixels >
+            0.9 * _productsScrollController.position.maxScrollExtent) {
           setState(() {
             _searchInProgress = true;
           });
@@ -165,12 +166,14 @@ class _ProductsScreenState extends State<ProductsScreen> {
             color: Theme.of(context).accentColor,
           ),
           onPressed: () {
-            showDialog(context: context, builder: _filterDialogBuilder).then((priceRange) {
+            showDialog(context: context, builder: _filterDialogBuilder)
+                .then((priceRange) {
               if (priceRange == null) {
                 // Dialog was closed
                 return;
               }
-              if (_minPrice != priceRange['minPrice'] || _maxPrice != priceRange['maxPrice']) {
+              if (_minPrice != priceRange['minPrice'] ||
+                  _maxPrice != priceRange['maxPrice']) {
                 setState(() {
                   _minPrice = priceRange['minPrice'];
                   _maxPrice = priceRange['maxPrice'];
@@ -202,11 +205,14 @@ class _ProductsScreenState extends State<ProductsScreen> {
     );
   }
 
-  PopupMenuItem<SortStrategy> _buildMenuItem(SortStrategy value, String text) => PopupMenuItem<SortStrategy>(
+  PopupMenuItem<SortStrategy> _buildMenuItem(SortStrategy value, String text) =>
+      PopupMenuItem<SortStrategy>(
         value: value,
         child: Text(text),
         textStyle: TextStyle(
-          color: _sortStrategy == value ? Theme.of(context).primaryColor : Theme.of(context).textTheme.bodyText2.color,
+          color: _sortStrategy == value
+              ? Theme.of(context).primaryColor
+              : Theme.of(context).textTheme.bodyText2.color,
         ),
       );
 
@@ -265,7 +271,9 @@ class _ProductsScreenState extends State<ProductsScreen> {
     );
   }
 
-  Widget _getProductCard(index) => _products.length - 1 >= index ? ProductCard(product: _products[index]) : Container();
+  Widget _getProductCard(index) => _products.length - 1 >= index
+      ? ProductCard(product: _products[index])
+      : Container();
 
   @override
   void dispose() {
@@ -390,7 +398,7 @@ class _ProductsScreenState extends State<ProductsScreen> {
               ),
               ButtonBar(
                 children: [
-                  FlatButton(
+                  TextButton(
                     child: Text(tr('apply')),
                     onPressed: () {
                       if (_priceFormKey.currentState.validate()) {
