@@ -14,7 +14,7 @@ import 'package:charity_discount/ui/programs/reviews.dart';
 import 'package:charity_discount/util/tools.dart';
 import 'package:charity_discount/ui/app/util.dart';
 import 'package:easy_localization/easy_localization.dart';
-import 'package:flushbar/flushbar.dart';
+import 'package:another_flushbar/flushbar.dart';
 import 'package:flutter/material.dart';
 import 'package:charity_discount/models/program.dart' as models;
 import 'package:charity_discount/services/affiliate.dart';
@@ -48,10 +48,11 @@ class _ShopDetailsState extends State<ShopDetails> {
 
     final appState = AppModel.of(context);
 
-    double sectionTitleSize = Theme.of(context).textTheme.headline5.fontSize * 0.7;
+    double sectionTitleSize =
+        Theme.of(context).textTheme.headline5.fontSize * 0.7;
 
-    final reviewsFuture =
-        _reviewsMemoizer.runOnce(() => locator<ShopsService>().getProgramRating(widget.program.uniqueCode));
+    final reviewsFuture = _reviewsMemoizer.runOnce(() =>
+        locator<ShopsService>().getProgramRating(widget.program.uniqueCode));
     Widget ratingBuilder = FutureBuilder<List<Review>>(
       future: reviewsFuture,
       builder: (context, snapshot) {
@@ -64,7 +65,8 @@ class _ShopDetailsState extends State<ShopDetails> {
           return SliverToBoxAdapter(child: loading);
         }
 
-        final titleColor = snapshot.data.isEmpty ? Colors.grey.shade500 : Colors.grey.shade800;
+        final titleColor =
+            snapshot.data.isEmpty ? Colors.grey.shade500 : Colors.grey.shade800;
         final reviewsTitle = Text(
           tr('review.reviews'),
           textAlign: TextAlign.start,
@@ -90,7 +92,10 @@ class _ShopDetailsState extends State<ShopDetails> {
 
         snapshot.data.sort((r1, r2) => r2.createdAt.compareTo(r1.createdAt));
 
-        List<Widget> reviewsWidgets = snapshot.data.take(3).map((rating) => RatingWidget(rating: rating)).toList();
+        List<Widget> reviewsWidgets = snapshot.data
+            .take(3)
+            .map((rating) => RatingWidget(rating: rating))
+            .toList();
 
         List<Widget> reviewSection = [
           Container(
@@ -111,7 +116,7 @@ class _ShopDetailsState extends State<ShopDetails> {
         if (snapshot.data.length > 3) {
           reviewSection.add(
             Center(
-              child: FlatButton(
+              child: TextButton(
                 child: Text(tr('seeAll')),
                 onPressed: () {
                   Navigator.push(
@@ -161,7 +166,8 @@ class _ShopDetailsState extends State<ShopDetails> {
         }
 
         var promotions = snapshot.data;
-        final titleColor = promotions.isEmpty ? Colors.grey.shade500 : Colors.grey.shade800;
+        final titleColor =
+            promotions.isEmpty ? Colors.grey.shade500 : Colors.grey.shade800;
         final promotionsTitle = Text(
           tr('promotion.promotions'),
           style: TextStyle(
@@ -186,8 +192,10 @@ class _ShopDetailsState extends State<ShopDetails> {
       },
     );
 
-    final productsFuture = _productsMemoizer.runOnce(() => Future.delayed(Duration(milliseconds: 500))
-        .then((_) => AppModel.of(context).getProductsForProgram(program: widget.program)));
+    final productsFuture = _productsMemoizer.runOnce(() =>
+        Future.delayed(Duration(milliseconds: 500)).then((_) =>
+            AppModel.of(context)
+                .getProductsForProgram(program: widget.program)));
 
     return Scaffold(
       appBar: AppBar(
@@ -266,7 +274,8 @@ class _ShopDetailsState extends State<ShopDetails> {
     );
   }
 
-  ClipOval _buildProvideRating(bool alreadyReviewed, BuildContext context, Review thisUserReview) {
+  ClipOval _buildProvideRating(
+      bool alreadyReviewed, BuildContext context, Review thisUserReview) {
     return ClipOval(
       child: Container(
         color: Colors.green,
@@ -357,7 +366,10 @@ class _ShopDetailsState extends State<ShopDetails> {
             padding: const EdgeInsets.symmetric(horizontal: 8.0),
             child: Text(
               getProgramCommission(widget.program),
-              style: Theme.of(context).textTheme.headline5.copyWith(color: Colors.green),
+              style: Theme.of(context)
+                  .textTheme
+                  .headline5
+                  .copyWith(color: Colors.green),
             ),
           ),
         ],
@@ -391,7 +403,9 @@ class _ShopDetailsState extends State<ShopDetails> {
     return Center(
       child: Tooltip(
         showDuration: Duration(seconds: 5),
-        message: widget.program.sellingCountries.map((country) => country.name).join(', '),
+        message: widget.program.sellingCountries
+            .map((country) => country.name)
+            .join(', '),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
@@ -400,7 +414,9 @@ class _ShopDetailsState extends State<ShopDetails> {
               style: Theme.of(context).textTheme.caption,
             ),
             Text(
-              widget.program.sellingCountries.map((country) => country.name).join(', '),
+              widget.program.sellingCountries
+                  .map((country) => country.name)
+                  .join(', '),
               style: Theme.of(context).textTheme.subtitle1,
               overflow: TextOverflow.ellipsis,
             ),
