@@ -6,7 +6,7 @@ import 'package:charity_discount/ui/app/util.dart';
 import 'package:charity_discount/util/amounts.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:easy_localization/easy_localization.dart';
-import 'package:flushbar/flushbar.dart';
+import 'package:another_flushbar/flushbar.dart';
 import 'package:flutter/material.dart';
 import 'package:charity_discount/services/charity.dart';
 import 'package:charity_discount/models/charity.dart';
@@ -150,7 +150,7 @@ class _DonateWidgetState extends State<DonateWidget> {
             ),
             ButtonBar(
               children: [
-                FlatButton(
+                TextButton(
                   child: Row(
                     children: <Widget>[
                       Icon(
